added navigation home, login, register, moved login register into auth folder, added global css
